Don't rewrite Content-type header in all HTTP requests
HttpInterceptor currently rewrites headers in all requests and sets
Content-Type to application/json. However, there are usecases when the
application needs to pass its own headers or keep the different
Content-type.

This commit changes HttpInterceptor to keep the original headers,
add what it needs (authentication) and if not set 'specific-content-type'
header, it sets the Content-Type to application/json.