Fix issue #14
The problem was caused by a conversion from JSON to dictionary. In the
end, all JSON values are strings. When original python code checked for
instance of int, it always evaluated to False, anything evaluated as a
string and since there was not a test for role == '0' or similar tests,
it was impossible to properly set roles. By trying to cast the value to
int and test for strings when that fails, we can fix the bug.