frontend FEATURE generate colorTheme SCSS based on app.config.json
Allow application to modify frontend colors via app.config.json
colorTheme record. The SCSS file is generated from the default template,
specified colors are replaced with a new value and the file is provided
to all the modules as _colors.scss file in the frontend top level directory.
This process is covered by bootstrap.py