gui CHANGE unify default host to connect frontend with backend
Default listening host 'localhost' for the backend can resolve to IPv4
as well as IPv6 address. On the other side, frontend was set to connect
to ::1 (IPv6 localhost) so in case the backend started (for any reason)
to listen on 127.0.0.1, frontend was not able to connect to it.

So we unify the default host to 127.0.0.1 which should work even in
environment where IPv6 is disabled/not supported.